pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-research"
version = "0.2.0"
description = "Autonomous research workflow framework with LLM-driven spec clarification"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
agentic-research = "agentic_research.cli:app"

[tool.setuptools.packages.find]
include = ["agentic_research*"]

[tool.setuptools.package-data]
"agentic_research" = ["templates/*"]
